+++ server/src/main/java/io/foodapp/server/dtos/Staff/SalaryHistoryResponse.java
@@ -1,5 +1,7 @@
 package io.foodapp.server.dtos.Staff;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class SalaryHistoryResponse {
     private int month;
     private int year;
-    private double currentSalary;
+    private BigDecimal currentSalary;
 }
